Set up config for jest/enzyme testing
