+++ babel.config.js
@@ -1,3 +1,7 @@
+// module.exports = {
+//   presets: ['@babel/preset-env', '@babel/preset-react'],
+// };
+
 module.exports = {
   presets: [
     [
