+++ client/src/index.jsx
@@ -1,26 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-
-class App extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      isLoad: false,
-    };
-  }
-
-  render() {
-    const { isLoad } = this.state;
-    if (isLoad) {
-      return (
-        <div>checker</div>
-      );
-    }
-    return (
-      <div>Loading</div>
-    );
-  }
-}
+import App from './components/app';
 
 ReactDOM.render(<App />, document.getElementById('app'));
