+++ webpack.config.js
@@ -20,4 +20,7 @@ module.exports = {
       },
     ],
   },
+  resolve: {
+    extensions: ['.js', '.json', '.jsx'],
+  },
 };
